Adds tests, up migration
